+++ plazoleta-service/src/main/java/com/pragma/plazoleta/domain/api/IDishServicePort.java
@@ -6,10 +6,15 @@ import java.util.List;
 
 public interface IDishServicePort {
     void saveDish(Long id_proprietary, DishModel dish);
-    DishModel findDishById(int id);
+
+    DishModel findDishById(Long id);
+
     List<DishModel> getAllDishes();
-    void updateDish(Long id_proprietary, int id_dish, DishModel dishModel);
-    DishModel updateActive(boolean active, Long idProprietary, int idDish);
+
+    void updateDish(Long id_proprietary, Long id_dish, DishModel dishModel);
+
+    DishModel updateActive(boolean active, Long idProprietary, Long idDish);
+
     List<DishModel> getDishesByRestaurantWithPaginationByCategory(Long idRestaurant, int page, int size);
 
 }
