+++ plazoleta-service/src/main/java/com/pragma/plazoleta/domain/spi/persistence/IDishPersistencePort.java
@@ -6,7 +6,9 @@ import java.util.List;
 
 public interface IDishPersistencePort {
     DishModel saveDish(DishModel dish);
-    DishModel findDishById(int id);
+
+    DishModel findDishById(Long id);
+
     List<DishModel> getAllDishes();
     DishModel updateDish(DishModel dishModel);
     DishModel updateActive(DishModel dishModel);
